Create Age Calculator
This program displays two input boxes using HTML & CSS  for the user to enter in their current age and a year. There is a JavaScript function that will tell the user the age they will be or where in the given year.  The information is displayed in the HTML page.